+++ packages/client-library-otel/src/utils/attributes/body.ts
