+++ packages/client-library-otel/src/utils/attributes/clone.ts
