+++ packages/client-library-otel/src/utils/attributes/headers.test.ts
