+++ packages/client-library-otel/src/utils/attributes/index.ts
