+++ packages/client-library-otel/src/utils/attributes/response.ts
