+++ packages/client-library-otel/src/utils/attributes/url.ts
